requirements.txt
python-docx==1.1.2
openai==1.12.0
python-dotenv==1.0.1
httpx==0.27.0
langdetect==1.0.9
PyYAML==6.0.1

